requirements.txt
streamlit
pandas
google-generativeai>=0.7.0
requests
python-dotenv